Do sprawdzenia.

+++ src/main/java/com/PLCompanyAccountingBackend/models/BusinessContractors.java
@@ -2,14 +2,12 @@ package com.PLCompanyAccountingBackend.models;
 
 import jakarta.persistence.*;
 
-import java.util.Date;
-
 @Entity
 @Table(name = "BusinessContractors")
 public class BusinessContractors {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "Business/Contractors name - C.4", nullable = false)
     private String businessContractorsName;
@@ -25,7 +23,7 @@ public class BusinessContractors {
         this.businessContractorsAddress = businessContractorsAddress;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
